primefield

Macro field_op

macro_rules! field_op {
    ($fe:path, $op:tt, $func:ident, $inner_func:ident) => { ... };
}

Emit a core::ops trait wrapper for an inherent method which is expected to be provided by a backend arithmetic implementation (e.g. fiat-crypto)