primefield

Macro monty_field_arithmetic

macro_rules! monty_field_arithmetic {
    (
        name: $fe:tt,
        params: $params:ty,
        uint: $uint:ty
    ) => { ... };
}

Add const fn methods to the given field element for performing field arithmetic operations, e.g. add, double, sub, multiply, neg.

This macro wraps a generic field implementation provided by the crypto-bigint crate, which is exposed as the primefield::MontyFieldElement type.