primefield

Module monty

Field elements which use an internal Montgomery form representation, implemented using crypto-bigint’s MontyForm.

Macros

monty_field_op 🔒

Emit a core::ops trait wrapper for an inherent method.

Structs

MontyFieldElement

Field element type which uses an internal Montgomery form representation.

Traits

MontyFieldParams

Extension trait for defining additional field parameters beyond the ones provided by [ConstMontyPrimeParams].

Functions

compute_t

Compute t = (modulus - 1) >> S

Type Aliases

MontyFieldBytes

Serialized representation of a field element.