primefield::monty

Trait MontyFieldParams

pub trait MontyFieldParams<const LIMBS: usize>: ConstPrimeMontyParams<LIMBS> {
    type ByteSize: ArraySize;

    const BYTE_ORDER: ByteOrder;
    const MODULUS_HEX: &'static str;
    const T: Uint<LIMBS>;
}

Extension trait for defining additional field parameters beyond the ones provided by [ConstMontyPrimeParams].

Required Associated Constants

const BYTE_ORDER: ByteOrder

Byte order to use when serializing a field element as byte.

const MODULUS_HEX: &'static str

Field modulus as a hexadecimal string.

const T: Uint<LIMBS>

T = (modulus - 1) >> S, where S = (modulus - 1).trailing_zeros()

Required Associated Types

type ByteSize: ArraySize

Size of a field element when serialized as bytes.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.