primefield

Struct MontyFieldElement

pub struct MontyFieldElement<MOD, const LIMBS: usize>
where
    MOD: MontyFieldParams<LIMBS>,
{
    inner: ConstMontyForm<MOD, LIMBS>,
}

Field element type which uses an internal Montgomery form representation.

Fields
inner: ConstMontyForm<MOD, LIMBS>
Implementations

impl<MOD, const LIMBS: usize> MontyFieldElement<MOD, LIMBS>
where MOD: MontyFieldParams<LIMBS>,

pub const ZERO: Self

Zero element (additive identity).

pub const ONE: Self

Multiplicative identity.

pub const LIMBS: usize = LIMBS

Number of limbs used by the internal integer representation.

pub fn from_bytes(repr: &MontyFieldBytes<MOD, LIMBS>) -> CtOption<Self>
where Uint<LIMBS>: ArrayEncoding,

Decode field element from a canonical bytestring representation.

pub fn from_slice(slice: &[u8]) -> Option<Self>
where Uint<LIMBS>: ArrayEncoding,

Decode field element from a canonical byte slice.

Slice is expected to be zero padded to the expected byte size.

pub const fn from_hex_vartime(hex: &str) -> Self

Decode a field element from hex-encoded bytes.

This is primarily intended for defining constants using hex literals.

Panics

When hex is malformed

When input is the wrong length

If input overflows the modulus

pub const fn from_uint_reduced(uint: &Uint<LIMBS>) -> Self

Convert Uint into MontyFieldElement, first converting it into Montgomery form:

w * R^2 * R^-1 mod p = wR mod p

Reduces the input modulo p.

pub fn from_uint(uint: &Uint<LIMBS>) -> CtOption<Self>

Convert Uint into MontyFieldElement, first converting it into Montgomery form:

w * R^2 * R^-1 mod p = wR mod p

Returns

The CtOption equivalent of None if the input overflows the modulus.

pub const fn from_u32(w: u32) -> Self

Convert a u32 into a MontyFieldElement.

Panics

If the modulus is 32-bits or smaller.

pub const fn from_u64(w: u64) -> Self

Convert a u64 into a MontyFieldElement.

Panics

If the modulus is 64-bits or smaller.

pub const fn from_montgomery(uint: Uint<LIMBS>) -> Self

Create MontyFieldElement from a Uint which is already in Montgomery form.

⚠️ Warning

This value is expected to be in Montgomery form and reduced. Failure to maintain these invariants will lead to miscomputation and potential security issues!

pub const fn from_montgomery_words(words: [Word; LIMBS]) -> Self

Helper function to construct MontyFieldElement from words in Montgomery form.

pub const fn as_montgomery(&self) -> &Uint<LIMBS>

Borrow the inner Uint type which is in Montgomery form.

⚠️ Warning

Make sure you are actually expecting a value in Montgomery form! This is not the correct function for converting out of Montgomery form: that would be MontyFieldElement::to_canonical.

pub const fn to_montgomery_words(&self) -> [Word; LIMBS]

Retrieve the Montgomery form representation as an array of Words.

pub fn to_bytes(self) -> MontyFieldBytes<MOD, LIMBS>
where MOD: MontyFieldParams<LIMBS>, Uint<LIMBS>: ArrayEncoding,

Returns the bytestring encoding of this field element.

pub fn is_odd(&self) -> Choice

Determine if this field element is odd: self mod 2 == 1.

Returns

If odd, return Choice(1). Otherwise, return Choice(0).

pub fn is_even(&self) -> Choice

Determine if this field element is even: self mod 2 == 0.

Returns

If even, return Choice(1). Otherwise, return Choice(0).

pub fn is_zero(&self) -> Choice

Determine if this field element is zero.

Returns

If zero, return Choice(1). Otherwise, return Choice(0).

pub const fn to_canonical(self) -> Uint<LIMBS>

Translate field element out of the Montgomery domain, returning a Uint in canonical form.

pub const fn add(&self, rhs: &Self) -> Self

Add elements.

pub const fn double(&self) -> Self

Double element (add it to itself).

pub const fn sub(&self, rhs: &Self) -> Self

Subtract elements.

pub const fn multiply(&self, rhs: &Self) -> Self

Multiply elements.

pub const fn neg(&self) -> Self

Negate element.

pub const fn square(&self) -> Self

Compute modular square.

pub fn invert(&self) -> CtOption<Self>

Compute field inversion: 1 / self.

pub fn invert_vartime(&self) -> CtOption<Self>

Compute field inversion: 1 / self in variable-time.

pub const fn const_invert(&self) -> Self

Compute field inversion as a const fn. Panics if self is zero.

This is mainly intended for inverting constants at compile time.

pub const fn pow_vartime<const RHS_LIMBS: usize>( &self, exp: &Uint<RHS_LIMBS>, ) -> Self

Returns self^exp, where exp is a little-endian integer exponent.

This operation is variable time with respect to the exponent exp.

If exp is fixed, this operation is constant time. Note that exp will still be branched upon and should NOT be a secret.

pub const fn sqn_vartime(&self, n: usize) -> Self

Returns self^(2^n) mod p.

This operation is variable time with respect to the exponent n.

If the exponent is fixed, this operation is constant time.