Crate primefield

RustCrypto: Generic Prime Fields

Generic implementation of prime fields built on crypto-bigint, along with macros for writing field element newtypes including ones with formally verified arithmetic using fiat-crypto.

Documentation

About

A prime field is a finite field of order 𝑝, where 𝑝 is a prime number. Because 𝑝 is prime, every non-zero element of the field has a modular inverse.

Prime fields are notable for their use in cryptography, particularly for their use as coordinates (a.k.a. base field) and the scalar field of elliptic curve implementations.

The implementation provided by this crate is built on crypto_bigint::modular, which provides a generic implementation of modular arithmetic with a modulus fixed at compile-time.

Minimum Supported Rust Version (MSRV) Policy

MSRV increases are not considered breaking changes and can happen in patch releases.

The crate MSRV accounts for all supported targets and crate feature combinations, excluding explicitly unstable features.

License

Licensed under either of:

Apache License, Version 2.0

MIT license

at your option.

Contribution

Unless you explicitly state otherwise, any contribution intentionally submitted for inclusion in the work by you, as defined in the Apache-2.0 license, shall be dual licensed as above, without any additional terms or conditions.

Re-exports

pub use bigint;

pub use bigint::hybrid_array as array;

pub use common;

pub use ff;

pub use rand_core;

pub use subtle;

pub use zeroize;

Modules

consts

Macros

bench_field

Write a series of criterion-based benchmarks for a field implementation.

fiat_bernstein_yang_invert

Emit wrapper function for a fiat-crypto generated implementation of the Bernstein-Yang (a.k.a. safegcd) modular inversion algorithm.

fiat_monty_field_arithmetic

Add fiat-crypto synthesized arithmetic impls to the given field element.

field_op

Emit a core::ops trait wrapper for an inherent method which is expected to be provided by a backend arithmetic implementation (e.g. fiat-crypto)

monty_field_arithmetic

Add const fn methods to the given field element for performing field arithmetic operations, e.g. add, double, sub, multiply, neg.

monty_field_element

Implements a field element type whose internal representation is in Montgomery form, providing a combination of trait impls and inherent impls which are const fn where possible.

monty_field_params

Creates a ZST representing the Montgomery parameters for a given field modulus.

monty_field_reduce

Write Reduce impls for a particular field implementation which delegate to MontyFieldElement.

test_fiat_monty_field_arithmetic

Test that the wrapper generated for a word-by-word-montgomery fiat-crypto implementation matches the internal representation used by MontyField*, e.g. do they have the same saturated modulus representation and agree on the value of R.

test_field_identity

Implement field element identity tests.

test_field_invert

Implement field element inversion tests.

test_field_sqrt

Implement field element square root tests.

test_primefield

Implement all tests for a type which impls the PrimeField trait.

test_primefield_constants

Implement tests for constants defined by the PrimeField trait.

Structs

Error

Error type.

MontyFieldElement

Field element type which uses an internal Montgomery form representation.

Enums

ByteOrder

Byte order used when encoding/decoding field elements as bytestrings.

Traits

FieldExt

Extension trait for ff::Field, intended as a place to put optimizable arithmetic operations.

MontyFieldParams

Extension trait for defining additional field parameters beyond the ones provided by [ConstMontyPrimeParams].

PrimeFieldExt

Extension trait for ff::PrimeField which enables specifying the endianness in which ff::PrimeField::Repr is encoded.

Functions

compute_t

Compute t = (modulus - 1) >> S

Type Aliases

MontyFieldBytes

Serialized representation of a field element.

Result

Result type.